Refactor Perron::Data to Perron::DataSource
Rename Data to DataSource to better reflect its purpose as a source of
data records rather than the data itself. This clarifies the distinction
between the data source abstraction and the individual data items it contains.
