+++ lib/perron/resource/sourceable.rb
@@ -40,7 +40,7 @@ module Perron
         end
 
         def combinations
-          datasets = source_names.map { Perron::Data.new(it.to_s) }
+          datasets = source_names.map { Perron::DataSource.new(it.to_s) }
 
           datasets.first.product(*datasets[1..])
         end
@@ -72,7 +72,7 @@ module Perron
             singular_name = name.to_s.singularize
             identifier = frontmatter["#{singular_name}_#{primary_key}"]
 
-            hash[name] = Perron::Data.new(name.to_s).find { it.public_send(primary_key).to_s == identifier.to_s }
+            hash[name] = Perron::DataSource.new(name.to_s).find { it.public_send(primary_key).to_s == identifier.to_s }
           end
 
           Source.new(data)
